tweak word monitor; comment #2; resolve "Exceptional ." in MachineReader
word monitor now only shows if the mouseY is on its latitude or when the
buttons are showing (mouseY on their latitude). also tried using the
'info' font (as for buttons. (NB: textSize() only affects the font after
font is loaded/set and does not play nice, visually, with fonts anyway.)
  on #2: related unwanted behavior: recreate/observe this by selecting a
Fluent Simple Spawner before it will shortly cause a page-turn. Note
that the first word of the new grid does not appear in the word monitor
and that VBs for this word do not run as we would like them to: with
distinct highlighting of the word and fade back to grid template
defaults according to the configuration of VBs.