resolved #14; extensive refactoring; comment #2
the fading mysteries turned out to be a very gnarly problem with
Mesostic Readers (their selectNext changes their members and shouldn't),
this is kludged for now. // The only problem with #2 intial reader
position is now that humans don't get to see good enterWord highlighting
of the initial word (when either the sketch loads or a new text is
selected).