working on application export - via p5
dch: early in the year you developed a way to generate a p5 version of
the sketch (you did this for Poetic Caption) that could use the p5
"Export Application ..." to make the platform-ready versions.
  I've edited this for RELC3 and I've also updated the p5 core.jar
because this was needed.
  It *almost* works, but there is a system-level problem which means
that the Mac app does not run. If you run one of the Linux apps, you'll
see what I think is producing the error. I'm adding this as a issue.